requirements.txt
Flask
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests